Test and pom update - Spring

+++ backend/habitStats_Server/src/test/java/pte/mik/habitstatsserver/webtest/GoalControllerTest.java
@@ -7,15 +7,12 @@ import pte.mik.habitstatsserver.controller.stat.GoalController;
 import pte.mik.habitstatsserver.dto.CreateGoalDto;
 import pte.mik.habitstatsserver.entity.stat.Goal;
 import pte.mik.habitstatsserver.entity.stat.Stat;
-import pte.mik.habitstatsserver.entity.stat.StatCategory;
-import pte.mik.habitstatsserver.entity.stat.UnitType;
 import pte.mik.habitstatsserver.repository.stat.GoalRepository;
 import pte.mik.habitstatsserver.repository.stat.StatRepository;
 
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -36,20 +33,20 @@ public class GoalControllerTest {
 
     @Test
     @Transactional
-    public void create() throws Exception{
-        CreateGoalDto goalDto = new CreateGoalDto("Test title", 99f, "<=", false,new ArrayList<>(latestStat().getId()));
+    public void create() throws Exception {
+        CreateGoalDto goalDto = new CreateGoalDto("Test title", 99f, "<=", false, new ArrayList<>(Math.toIntExact(latestStat().getId())));
         assertThat(controller.create(goalDto)).isEqualTo("Done!");
     }
 
     @Test
     public void read() throws Exception{
-        assertThat(controller.listAll().stream().map(Goal::getId).filter(integer -> integer.equals(latestGoal().getId())).findFirst().isPresent());
+        assertThat(controller.listAll().stream().map(Goal::getId).filter(Long -> Long.equals(latestGoal().getId())).findFirst().isPresent());
     }
 
     @Test
     @Transactional
     public void delete() throws Exception{
-        Integer idActedOn = latestGoal().getId();
+        Long idActedOn = latestGoal().getId();
         assertThat(controller.delete(idActedOn)).isEqualTo("Done!");
     }
